Tests: fix anonymous user creationn
The recent django-guardian upgrade made an assumption about the passed
in arguments visisble: the database alias has to be passed in for the
create_anonymous_user() signal handler. Providing it fixes the tests,
but ultimately it would be better if we replace this function with our
own logic to not misuse the guardian internal signal handler. To make
this easier, the currently used logic has been factored into a separate
fuction. In the future, we might want to do something similar to this:

https://github.com/django-guardian/django-guardian/blob/devel/guardian/management/__init__.py#L23