Web client: use flexbox for top tool bar
This makes the HTML order of toolbar elements more natural and it allows
for a saner wrapping mode. If there is now not enough horizontal space
no wrapping is performed, but the "data" toolbox (the one showing all
widget buttons) shrinks. Only elements that fit the available width are
displayed there now. Invisible items are not a problem, because all
widgets can be reached through the "Open Widget" dialog or Ctrl + Space.

Also the buttons displayed in the data toolbar element are now recreated
with every tool change. This makes it possible to have individual tools
provide their actions for this toolbar without requiring additional
toolbar items (like it is currently the case for the ontology widget)
and allows the removal of some explicit visibility control of the data
toolbar element in the Tracing Tool.

A flexbox layout is also applied to the second tool-specific toolbar.
The commonly visible Z and zoom sliders are shrink now when necessary
and as long as reasonable.