Node filters: fix application of result node mapping function
Node filters can apply a mapping to the result node collection. This is
useful when additional information like the node location should be
available right away to the result handler. For instance, the Volume
Manager uses this functionanlity to obtain node locations to generate a
preview mesh in a 3D viewer. The application of this mapping functionn
was broken since a1110de439 (June 2019) and with it the possibility to
create convex hulls and alpha shapes. This is fixed now.

Fixes catmaid/CATMAID#1961