Web client: add layout option to URL
A CATMAID deep-link URL can now include the layout that should be
displayed using the `layout` parameter. It takes a layout specification
as it is used in the Settings Widget and the Layout menu.

For instance, to load a Neuron Search for annotation "papers", the
following URL segment could be used:

  layout=h(XY, { type: "neuron-search", id: "neuron-search-1", options: {"annotation-name": "papers"}}, 0.6)