Upgrade channels from v1.1.8 to v2.4.0
This newer version of channels supports Django 2 and Django 3. It
required some rework of the ASGI implementation in the back-end.

Note that this version doesn't come with an IPC shared memory channels
layer by default. If inter-process sharing is wanted/needed (which is
typically the case with multiple back-end workers), channels_rabbitmq or
channels_redis needs to be used. The documentation has information about
this.