Upgrade Django from v2.0.9 to v3.0.2
With respect to general Django 3 changes, only little adaption had to
happen in our code:

- The context arguments from the `from_db_value()` method in custom
  model fields has been removed. We didn't use it anyway.
- The `django.shortcuts.render_to_response()` was removed in favor of
  `django.shortcuts.render()`, which takes the same arguments, but also
  expects a request.
- The jsonfield dependency was dropped entirely, because the lack of
  support for Django 3 and we effectively didn't use it. We only used it
  still for the first database migration, but this could be replaced
  with the Postgres native jsonb type, which we migrated to as one of
  the next things after the initial migration anyway. This doesn't pose
  a problem for those later migrations.
- Channels had to be upgraded to the most recent version. The version we
  used wasn't compatible with Django 3, and probably not even with
  Django 2 entirely.
- The `admin_static` template tag library has been removed in favor of
  the `static` library, which works the same.

The main driver behind this update is that it allows us to return more
than one field from an INSERT SQL query that has been issued by the ORM.
An indirect consequence of this is that we can't use the default values
provided by the database, because the ORM infrastructure wasn't able to
create an ORM object that has all it's fields filled either by an
explicitly provided value or default values that would have been
returned from the database (if the database is the one who should
provide them). A more detailed description on why this is a problem,
copied from catmaid/CATMAID#1949 and how Django 3 helps here:

Luckily one commit made it in just before feature freeze that allows
Django ORM querysets to return multiple values from database queries
(see django/django#9983). On object creation, Django so far returned
only the ID (`RETURNING id` at the end of an `INSERT`), which prevented
the use of database default values. This leads to inconsistent behavior
of the default timestamp values of creation_time and edition_time: if
created in raw SQL, the start of the transaction (now()) is used as
default timestamp. When created through the ORM Django's default value
is used (datetime.now), which is independent from the transaction. For
simple object creation, the Django ORM is used in quite a few places in
our code base.

While this doesn't sound like a big deal, it it is in fact a problem for
an assumption the history system has: CATMAID stores transaction
information in the catmaid_transaction_info table and each transaction
is identified by a transaction ID and a timestamp (the beginning of the
transaction). The transaction ID alone isn't enough, because transaction
wraparound (transaction ID overflow/reset) or backup restores can cause
duplicate IDs. Therefore the timestamp is stored as well. And the
assumption is that each tracked table row is identified by its creation
time and transaction ID. Therefore, for each object that is tracked by
CATMAID's history system, the transaction ID and a timestamp is stored
as well, and both the timestamp (creation_time for most tables).

If these timestamps don't match, then historic entries can't be
identified correctly. The time delta is very low though and I believe
this can be fixed programmatically. This also only becomes an issue if
one wants to look up historic objects and one has ID duplicates. Not
very likely, but not unrealistic: I came across this as part of a setup
restored from a backup.

Fixes catmaid/CATMAID#1949