Mark value field of client_data table as text field in initial migration
This is needed for Django 3, or otherwise the DatastoresApiTests fail
and newly inserted data is stored as text string rather than an object
in JSON format. It seems like Django 3 tries to be clever when applying
migrations and doesn't reapply column data type changes if it sees that
the data type did not change according to its list of applied
migrations.  This wasn't the case before Django 3 and therefore this
issue wasn't noticed.

I just wasted a whole day debugging this…