Graph widget: handle delete key properly
Pressing the Delete key with a node selected in an active Graph Widget
will now ask for the removal of the node from the graph, not anymore
whether the active node should be deleted completely.

The delete key needs to be handled in the handleKeyPress() function,
rather than the handleKeyUp() function, because otherwise the Tracing
Overlay will get the event.

Also, events that are handled by a particular handler will not be
attempted to be handled by other handlers.

Fixes catmaid/catmaid#1939