Graph widget: don't add arrowheads to all edges on refresh
This bug made arrowheads appear on undirected edges after a refresh of
the graph display. The main problem was that the "arrowshape" field of a
node would lead to added arrowhead information on edges on a reload (and
not on nodes).

Fixes catmaid/catmaid#1819 (at least the second issue, since the first
one didn't reappear and we don't have anything to go on).