Graph widget: fix splitting of node linked to prev. split node
So far the graph widget would raise an error if one tries to split a
node that is postsynaptic to to an already split node. It looks like
this change should fix it, but the complexity of the code makes it hard
to see every corner case.

See catmaid/catmaid#1819