Skeleton back-end: improve node-overview query construction
Let psycopg2 escape query parameters rather than formatting the query
string manually. Given that we cast the parameters to integers
beforehand, there was no risk for SQL injection, but the new version is
more consistent with the other queries and makes it easier to prevent
SQL injection in future changes.