Graph widget: fix storing the node size during an update
Recently node size information has been attached to individual nodes to
allow loading nodes with different sizes from GraphML files. Part of
this was to persist the per-node-width and -height during updates. This
process accidentally used node.size() rather than node.width() and
node.height(). The function used returns however the size of the node as
a node collection. This shrunk nodes after graph updates to the size of
1.