Tracing: add support for stable merge direction annotations
If a join partner in a skeleton merge is annotated with "stable", it
will now automatically win the join, i.e. its ID keeps on existing and
the other partner is merged into it, regardless of specified merge
order. If both merge partners are marked "stable", the merge is
canceled. This can be disabled by setting the project or instance
setting "Stable join annotation" in the Tracing section of the Settings
Widget to an empty string. This setting is only respected on the project
or instance level by the back-end. If instead of an empty string,
another annotation is specified, it will act as the new table join
annotation.