Merge dialog: show warning if merge winner is preselected through ann.
If there is a "stable join annotation" selected (by default "stable"),
then the merge dialog will check if any of the two merge partners are
annotated with it. If both are annotated a warning is shown that no
merge will be possible and if the smaller one is annotated, a warning
will be shown as well, noting which skeleton will win the merge.