Upgrade networkx from v1.11 to v2.4
This required quite a few changes, but luckily most of the functionality
has tests, which made most changes easy to verify. Both the v2 and the
v2.4 release promise performance improvements of features that we use.
Also, since v2.4 Python 3.8 is supported.

NetworkX is still much slower than networkkit and graph-tool. A future
switch to a competitor might still be a good idea.

See catmaid/catmaid#1611