64 bit concept ID migration: create indices concurrently
We require Postgres 11 and have the option to create indices
concurrently. This is very helpful for bigger CATMAID instances where
this migration could take many hours. Without concurrent index creation,
the migration requires exclusive locks and blocks all connection to the
database and no use is possible. With concurrent index creation CATMAID
can be used, but is just a bit slower until all indices are created.