Graph widget: make node width/height access more robust
This fixes the creation of groups, which so far don't have a node_width
and node_height property. The width and height for each node is now only
read from the node if it is actually available. Otherwise the default
size will be used.

Thanks to Margarida Brotos for reporting this issue.