WebSockets: make error handler more robust
The web socket error handler deletes any existing onclose handler of the
socket by setting it to null. This fails when there is no socket anymore
after the error. This seems to be possible to happen in rare cases
(according to the error log) and checking if there is a socket should
take care of this situation.