Tracing tool: don't try to show stack coordinates without active stack
In rare cases mouse move event handlers execute after a stack viewer has
been closed and removed as active stack viewer from the Tracing Tool. I
have seen this popping up in the error logs now and then. To not have
the coordinate printing handler raise an error in this case, the stack
coordinates are only included if there is an active stack viewer.