Cropping: fix computation of # slices to crop with stacks with translation
Carolyn Ott found this issue: when croping multi-slice subvolumes, the
computed number slices to crop was wrong. This was due to the computed
stack space Z coordinate was enforded to be in the range [0,
max-z-of-stack] without respecting the translation of this stack with
respect to the wold origin. This is fixed now by not enforcing stack
boundaries for computing the min/max stack space Z coordinates for
finding the number slices to crop.