Docs: fix watch in vagrant
By default, watchdog looks for file system events.
Files changed by the host do not emit events to the guest and vice
versa, so the the watcher would not trigger if it was running in the
guest and the files were changed by the host.
Watchdog has a "debug" polling mode which resolves this.