Graph widget: implement more complete node filtering
This fixes some issues with the previous implementation that resulted in
discrepancies between the Graph Widget and other widgets, like e.g. the
Connectivity Widget, with filters applied. The new implementation will
first obtain all treenodes allowed by the node filters and then finds
all connectors connected to them. These in turn are now provided to the
back-end, which now respects these connector node IDs as optional
additional filter. This makes the resulting values match other widgets
and allows use of additional subgraphs (e.g. axon/dendrite split), which
wasn't possible before.

In the front-end, the checkbox to disable node filters completely for
the Graph Widget moved from the Subgraphs tab to the Edges tab, because
this is what is ultimately filtered: edges/links. The "Circles"
expansion in the Grow tab respects node/edge filtered connector nodes as
well.

This also updates a few related SQL queries in the graph computation on
the back-end to make better use of psycopg2's parameterization rather
than build complete queries using string concatenation. This not only
helps against potential SQL injection attacks, but makes it often times
much more readable.

Additionally, the use of `IN` in some SQL queries has been replaced with
the use of `= ANY`, because it makes use of arrays easier, which can be
passed directly to psycopg2. Practically this results in the same plan.