Connectivity widget: make node filter application consistent with Graph Widget
A connection is now valid according to the selected filters if it uses a
connector node that is linked to a node allowed by the node filters.
Previously, the start node of a link had to be allowed explicitly. The
new version is more in line with user expectations. The node filters are
still only applied to the query skeletons. This makes the filtering
numbers consistent with the Graph Widget.