Connectivity widget: fix hiding too many rows w/ active node filters
If some situations with active node filters, a partner row was marked
wrongly to be ignored. This is fixed now and was originally discovered
by Lisa Marin. The problem was especially noticeable with multiple query
skeletons and active node filters. If the aggregated number of nodes
filtered was larger than the number in links for each query skeleton,
rows ended up being ignored. Adding confidence and synapse count
thresholds made this problem more likely to show up.