Graph widget: fix node width/height setting
The node width and node height settings fields in the Graph Widget's
properties dialog didn't have any effect anymore, a problem noticed by
Nusreen Imambocus. This is fixed now. The problem was that the place how
node width and node height are stored changed recently: instead of a
plain style property, it is a function that will give a node's data
object priority, should it have a node_width or node_height field.
Therefore, an update of node width or node height requires now an update
of a node's data object.

Plus, the widget scoped default fields for node width and node height
weren't updated.