Tracing tool: make sure active node centering is properly queued
When centering on the active node, the active node should only be
requested/read after all present tasks have been finished, because they
might affect the active node. For instance, when queuing node creation
operations on slower networks and wanting to center on a new node when
it isn't quite done yet, the past implementation would center on the
parent node, rather then the new node. This is fixed now.

Fixes catmaid/catmaid#1990