Export management command: fix duplicate tag export
Because tags are extracted from all links of tags to treenodes (to only
get used tags), duplicate tags are a consequence of both a tag being
used multiple times and using a list to collect the tags. To avoid this,
a set is now used to collect the tags. Thanks to Rob Court for bringing
this up!

Also, the log message is moved to before starting the actual work.