Tracing tool: make automatic cache update more robust
A wrong conditional prevented the refresh callback to be stopped after
the tracing tool was closed. If the whole project was closed and the
browser window was then open long enough (>60min), an error was raised
inside the cache refresh callback, because no project is available. This
is fixed now. Additionally, the cache refresh callback will only attempt
to do something meaningful if there is a project available.