Project/stack listing: fix inclusion of wrong duplicate stack groups
If a stack with a stack group was used in more than one project, wrong
duplicate stack groups were include in both the project listing and the
project export. This situation was so far not represented in the tests,
which is fixed by this commit as well.

Additionally, some manually created VALUES lists in involved SQL queries
have been replaced by passing lists to cursor.execute() and UNNEST,
which is much simpler and cleaner.

This also added to the slowness of the importer.