Stack viewer grid: override default overview visibility
The stack viewer grid is used in both the Treenode Viewer and the
Connector Viewer to show multiple nodes at the same time. In those mini
stack viewers the overview should be disabled even if the overview is
displayed by default in regular stack viewers.

This is made an optional constructor setting, because it is a central UI
component that is used differently in different contexts.