Connectivity widget: improve performance of cell invalidation
Some information in the partner tables is loaded dynamically, e.g. the
review information or neuron names are updated. This so far triggered
the invalidation and therefore recomputation of the complete rows with
updated information. It tuns our to be much faster to only update the
changed cells rather than the whole row. Especially for large tables
this can make a difference of many seconds. Thanks to Sanja Jasek for
pointing me to this problem with the Connectivity Widget.