Connectors: include unlinked connectors in bounding box query
The /{project_id}/connectors/in-bounding-box so far only returned
connectors that were linked to a treenode. It is however expected that
the function returns also unlinked connectors. This is fixed now an
optionally the only_linked flag can be passed in to return to only get
linked connectors. Thanks to @jtmaniatesselvin for pointing this bug
out!

See catmaid/catmaid#1999