Neuron navigator: fix sorting of annotation listings
The sorting was broken with the introduction of the ID column end of
last year. Ultimatately, the reason is the annotations/table-list API,
because it is is coupled so closely to DataTables spec and uses e.g.
numeric indices for defining which column to sort. We should replace
this some point with a more general API.