Neuron navigator: remember filter term in annotation list
This allows to keep a filter term if the currently displayed annotation
list has search/filter term defined for the table. Navigator nodes can
be refresh due to different events and actions and it is confusing if
the filter term disappears due to a refresh. Thanks to Sanja Jasek for
reporting this issue.