Skeleton analytics: fix error when adjacents = 0
The result of adjacents = 0 is that no no local graph for search for
connector duplicates is created. The relevant code however expected a
graph to be created. This problem wasn't really visible in networkx 1.x,
because the shortest path function would return before this point. With
networkx 2.4 this changed and the non-existent graph becomes an error.