Tracing overlay: immediately draw radius selections
Previously radius selection and measurement only drew after moving this
mouse. Since no indicator is rendered without mouse movement, this is
both confusing as to whether this mode has been entered, and tedious for
repeated radius annotation as it can require moving the mouse twice:
once to draw the indicator, then again to select the radius with the aid
of the indicator.

Instead, draw the surrounding circle immediately.