Add catmaid-synister as known extension
This is an extension developed with @nilsec and @funkey. In its first
iteration it will be used to look-up automatically generated
neurotransmitter information for the FAFB dataset (or any other that
such a dataset is generated for). This information will be attached to
connector nodes as tags. For now the plan is to have then consensus
look-ups on neurons in CATMAID to see the most common neurotransmitter
labels. Once this is working, tags for either all connectors or a subset
can be created. A database trigger will drop the existing
neurotransmiter tag and queue the look-up for a new look-up. This queue
is then consumed by an async worker.

To make development easier in the extension code, I add it already as
known extension here.