OAuth2 login: fix login menu links for subfolder setups
If CATMAID isn't run at the root of a domain, the OAuth2 login links
weren't working, because the subdomain part was added twice. Once by the
back-end and once by the front-end. This is fixed now and the front-end
just used the URLs directly that are provide by the back-end.