Improve display of 404 errors
Since a recent change the error message of a 404 error is just displayed
as "(undefined)", which doesn't really win a price for being helpful. If
the server now provides a non-HTML response text it is used, otherwise a
default error message is used that shows at least the unavailable
resource. The HTML test is done, because in debug mode the back-end will
return an HTML error page.

This also wraps 404 errors in a proper front-end type.