Pixi extensions: add simple fragment shader pipeline
This allows for manual control of min/max scaling via the console, and
builds towards future ability to perform shader computation before tiles
are rendered to a render buffer. Note that presently TypedSpriteRenderer
is still shaded between all TypedSprites of the same type, so this
pipeline is not granular enough for most intended uses. Consequently no
configuration is exposed for now.

Related to catmaid/CATMAID#1955.