Don't attempt to use Settings instances before initialization
Settings object scopes are only available after the async load() call
has returned. This isn't ideal, because it prevents synchronous use of
objects that require settings (e.g. the stack viewer). During runtime
this is normally no problem, but tests can break due to that. Still, we
might want to consider to initialize settings with their default values
and have a flag whether these settings are in sync with the back-end.