Tests: fix overlay update queuing tests
The queuing of the assert issuing function was wrong, because it was
executed as synchronous continuation of the finished submitter. It needs
to wait until other handlers are run, which can be done by queuing it
using a promise rather than a synchronous continuation.

This problem highlights unclear semantics of submit.then(), because it
it suggests that it is being run at a promise while it is infect not.
The queued function is injected into the submitter queue. To create a
proper promise that is independent of the submitter queue,
submit.promise() needs to be used.