Add min/max controls to typed image block layers
Allow for typed sprites to be associated with separate, owned renderer
plugins. This allows for the plugin for a sprite set, such as a tile
layer, to have its own shader parameters. This allows image blocks
layers to control their min/max mapping from their source type to the
float render buffer without interfering with other layers of the same
data type.