Stack: respect actual mirrors for default downsample factors
The default downsample factors for a stack without explicit factors
defined, changed slightly: if users relied on the automatic creation of
downsample factors for image stacks, they need to make sure the changed
behavior still works for them or enforce the previous configuration with
an explicit specification of downsample factors. Previously as many zoom
levels as needed were created to map the image size in a quadratic
fashion per level onto a 1024x1024px tile. Instead of hard-coding the
1024px, the minimum tile size defined in the mirror set is used. Without
any mirrors, no extra zoom levels will be assumed.

The docs have been updated accordingly and brought into line with the
current importer logic. The `zoomlevels` keyword is not respected
anymore.