Arbor: fix two bugs in connectedFractions()
This algorithm finds the largest connected treenode sets in a the passed
in node IDs. It does this by building up a tree from every passed in
node until no more valid nodes are found or a previously found tree is
hit. In this case the newer tree is merged into the existing one. The
current implementation keeps track of the current "active" tree and
during a merge this information is replaced with the merge target tree.
This was done correctly, but along with the tree there is also the root
node ID stored. It turned out it was updated incorrectly. At the point
the update happens "paren" is the top of the tree that is merged into
the other one. This is however not what should become the reference to
the new tip after the merge. This should be the top of the merge target
tree. And this is the first thing that this commit changes: The root
node reference of the active tree is now set correctly after a merge to
the merge target root.

Second, in case a tree is built up without ever hitting another tree,
the <arbors> indexing was wrong. It would still map the original root
node ID to the created arbor instance. This is wrong, because the index
should keep track of current valid trees, their root nodes mapped to
them. This is fixed now and the original index entry is replaced with
the actual root node being mapped to the new arbor.

This function is only used in one place, the Measurements Table, where
this problem also become visible. Thanks to Lydia Gruber for reporting
this issue!