Authentication: add GET API to retrieve token for anonymous user
The anonymous user is the representation of a user who isn't logged in
or authenticated. For remote data access, it would be useful if
anonymous access on the API level would be available not only for GET
request, but also POST requests. In order to do that, the API token of
the anonymous user needs to be made available. The front-end doesn't
expose the API key of the anonymous user anyway and probably doesn't
need to. Allowing access to the API key of the anonymous user through a
GET request will make adding remote CATMAID instances easier to add to
other CATMAID instances, given the remote servers provide anonymous
access. If they do, the other CATMAID instance can just use the URL
alone to add a remote CATMAID server, because it can retrieve the
anonymous API key by itself, using a GET request to avoid CSRF checks.