Graph widget: prevent additional edges in grouping w/ undirected links
Edge IDs between groups follow the pattern <group1>_<group2>. This is
used to group edges between two groups and with directed links, this
naming works well. In the case of undirected links, like desmosomes,
this can lead to accidentally created separate edges between groups,
depending on in which group the source and target node are. To prevent
this, group names are sorted for undirected edges before being combined
separated by an underscore. Thanks to Sanja Jasek for reporting this
problem!