Exporter mgmt. command: add option to export limited types of tags
With the help of the new --allowed-tags-only parameter it is possible to
limit the kinds of tags that are exported. Without this parameter, all
tags are exported. If --allowed-tags-only is provided without arguments,
a standard set of known tags is included (uncertain end, posterior end,
ends, anterior end, not a branch, really ends, uncertain continuation,
soma, microtubules end, out to nerve). Alternatively, such a list can be
provided as argument to this parameter.