Upgrade Three.js from v104 to v119
This makes us use the latest version of THREE.js with some performance
improvements relevant to us. In order to make this update work, I
followed the migration guide [1] and had to change the following things:

- The WebGLRenderer.context field has been replaced by a getContext()
  function.
- BufferGeometry.addAttribute() has been renamed to BufferGeometry.setAttribute().
- BufferGeometry.removeAttribute() has been renamed to BufferGeometry.deleteAttribute().
- Geometry.applyMatrix() has been renamed to Geometry.applyMatrix4().
- Object3D.applyMatrix() has been renamed to Object3D.applyMatrix4().
- Material.vertexColors is now a boolean. The default value is false.
- InstancedBufferGeometry.maxInstancedCount has been renamed to
  InstancedBufferGeometry.instanceCount.
- BufferAttribute.dynamic and BufferAttribute.setDynamic() have been
  deprecated. Please use BufferAttribute.usage and
  BufferAttribute.setUsage() instead.

A new small dependency (chevrotain) had to be added in order to keep the
VRMLLoader working, which we use to load volumes.

[1] https://github.com/mrdoob/three.js/wiki/Migration-Guide