Neuron search: add header row to CSV export
This is more consistent with other CSV export and apart from that needed
to make clear that the first column is a neuron ID. Otherwise it was
often mistaken for a skeleton ID. Thanks to Marta Costa for this
suggestion.

A future version could add a preview dialog and include also the option
to export skeleton IDs. This dialog could then also make the heaer
optional again, if that would be needed.

Fixes catmaid/catmaid#2019