Graph widget: fix error when ungrouping groups w/ non-synaptic members
At the moment non-synaptic groups have an ID that is an actual number.
To make sure we are dealing with strings when sorting source and target,
we now create string version of source and target explicitly. We might
want to make sure though that IDs are always strings.

Thanks to Sanja Jasek for reporting this issue!