Exporter: remove unneeded conditional
The test for "tag_links" isn't needed, because in all relevant cases,
"tags" represents the exact same information and is tested already.

Additionally, testing for "tag_links" leads to a use-before-assignment
error if the outer conditional is true (no tag setting overrides), but
no tags should be exported. This if fixed now as well.