New tile source: CloudVolume based tiles, generated by back-end
This new CloudVolumeTileSource uses the cloud-volume library in the
back-end to fetch image blocks and generate tiles for the front-end.
While slow for many users, it is a convenient way of make neuroglancer
volumes available to CATMAID. It works well for smaller teams and
smaller block sizes, it seems. The new tile source has ID 13.