Tracing tool: allow straight line distance measurements between nodes
Node distance measurements (icon in tracing tool bar) are now allowed
also between nodes of different skeletons. In this case only the
straight line distance will be displayed. Both the straight line
distance and the path distance will be displayed for nodes of the same
skeleton.

Thanks to Andrew Dacks for this suggestion.