Stack: data view thumbnails for H2N5 mirrors
This allows H2N5 mirrors to show a thumbnail tile in data views. N5
mirrors no longer return a valid overview URL, as these requests are
always broken. The overviews are not of the appropriate size (192px)
because CATMAID cannot parse the entire H2N5 URL format without knowing
metadata of the N5 container, but are instead simply one tile in size.

I dislike exposing details of tile sources to the backend, but it is not
unprecedented and a pragmatic fix for none of our now mostly H2N5 stacks
having thumbnails.