Neuron navigator: fix sorting problems in annotation table
Sorting of an annotation table in the Neuron Navigator widget is
currently done on the back-end through a rather complex API, implied by
Datatables. This API includes sorting based on an index number for a
column. This implicitly implies a column order, but since the function
that is calling this API is used in different column configurations, the
index number provided by the table isn't always what is assumed by the
back-end.

The proper fix would be to replace this API with something more readable
and get rid of index numbers. For now this workaround makes the columns
at least sortable again. If more time comes available the front-end
table API use is updated.

Fixes catmaid/catmaid#2026