Tracing tool: fix name service handling during unregister()
If a CATMAID Tool is unregistered, we can't assume the project is
destroyed with it. While this might be the case at times, just clicking
the Tracing Tool icon without closing the project will cause a call to
unregister(), too.

So far it was assumed that we can clear the Neuron Name Service for this
project und unset the active node. Unsetting the active node is merely
annoying, but clearing the NNS breaks the assumption for open widgets
that they can query names of neurons they are concerned about, leading
to a neuron name of "null".

This is fixed now and both the NNS and the Active Node will now listen
to the project destruction event explicitly, rather than being called by
the Tracing Tool's unregister() method. This improves the separation of
concerns and keeps the logic local to the rest of the code of the NNS
and the Active Node.

Fixes catmaid/catmaid#2036