Skeleton annotations: fix unsetting ATN on project destruction
We can only assume there is a project if there is a node passed in to
set(). If this isn't the case, for instance like during project
destruction, we can't assume this. We however also only need a project
reference if there is a node to be set. The code reflects this now as
well. Before, there was an error raised on project changes with an
active node selected in the Tracing Tool.

See catmaid/catmaid#2036