AJAX requests: ensure text response type on non-200 status
If a response did not return with status 200, we can suspect something
went wrong and therefore would expect text data even if binary data was
explicitly requests (e.g. tracing data, NRRD export, etc.). Without
forcing the response type to text in this case like we did so far, the
actual error message will be swallowed by the response parsing logic,
because the binary data is attempted to be parsed as JSON, due to the
response type being set to application/json for errors.