Split/merge dialog: fix warning message on merging stable skeletons
If a skeleton is marked as stable using the "stable" annotation (or
others that are configured), it will always be the winner of a merge and
if both partners are marked as stable, the merge is not performed.

The warning message shown to indicate the presence of a stable
annotation was wrongly worded. Thanks to @jasper-tms for reporting this
issue! It is fixed now.

Generally, the UI should indicate the stable skeleton more clearly and
not only in a warning message.

Fixes catmaid/catmaid#2039