Tracing overlay: fix wrong type for tracing data mirror API
The API used for tracing data mirrors is based on the remote server
configuration as configured through the Settings Widget. The settings
object stores is however not what should be used directly, like it was
before. Instead an API type is expected. This is fixed now. Thanks to
@ktedore for reporting this problem!

Fixes catmaid/catmaid#2035