Import/export widget: improve default name logic
So far, if no name was provided in the UI, the imported neuron would
just have no name. Now the user is asked whether the import should be
canceled or if the file name should be used as default name, if no name
was provided by the user.

Also, if multiple SWC files are imported and a name is provided, an
incrementing numeric suffic will be added to each name.