3D viewer: fix camera position change w/ mouse wheel
The camera can be moved (rather than the focal-length changed) when Alt
is pressed during the use of the mouse wheel. This was broken because
the logic to compute the location change was factored into its own
function, ignoring that a now internal value was used in further down.
This is fixed now by returning the required value from the factored
function.