Remote data import: fix import for data from local projects
If an import request doesn't come with a particular API instance, it is
assumed to come from the local instance. However, the import code
currently expects an explicit source URL reference. This is now done by
this commit by using the absolute URL of the current instance if nothing
else is provided.

Fixes catmaid/catmaid#2040