Skeleton API: keep original annotations around when updating skeleton through import
The skeleton import API can update the skeleton data for a particular
neuron ID or skeleton ID. If this is done, it should keep the existing
annotations around as a default. Therefore, annotations provided to the
import API will now be appended to any existing annotations, if a
skeleton ID or neuron ID is provided. Existing annotations can be
removed before adding new ones (if any), by setting the
replace_annotations parameter to true.

Fixes catmaid/catmaid#2042