Stack viewer: fix bug when removing copies of stack layers
If a stack layer is removed, CATMAID checks whether this was the last
stack layer for the stack ID of the removed layer. If this is the case,
then the stack is removed from the stack viewer. This check failed as it
used a forEach() call on a Map object like a filter() call on an Array.
This is fixed now.