Layer settings: add support for layer settings storage in datastores
Layer settings like opacity, blending more or filters can now be stored
as part of the regular settings and can be defined for the regular
scopes like in the Settings Widget (e.g. session or project). These
settings are stored per layer type and source reference (e.g. stack ID).
In order to save, restore and clear those settings, there are now three
new buttons at the bottom of each layer that supports settings storage.

There seem to be corner cases of these settings being overwritten if
WebGL can't be loaded for some reason, which should be prevented or
mitigated.